projetchat/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
argon2-cffi==23.1.0
Pillow>=10.3.0
python-decouple==3.8
drf-yasg==1.21.7
cryptography==41.0.7 